+++ SIL.DblBundle/DblMetadata.cs
@@ -29,9 +29,10 @@ namespace SIL.DblBundle
 		public int Revision { get; set; }
 
 		/// <summary>
-		/// Apparently it is possible to have revision set to "" in the XML. I'm not sure if this is new with version 2+ or if we just never saw it before.
-		/// Anyway, previously, Revision was simply an int as above, but I had to add this surrogate field to be able to handle the case when it is set to ""
-		/// while also not breaking the API.
+		/// Apparently it is possible to have revision set to "" in the XML. I'm not sure if this
+		/// is new with version 2+ or if we just never saw it before. Anyway, previously, Revision
+		/// was simply an int as above, but I had to add this surrogate field to be able to handle
+		/// the case when it is set to "" while also not breaking the API.
 		/// </summary>
 		[XmlAttribute("revision")]
 		public string Revision_Surrogate {
@@ -52,16 +53,22 @@ namespace SIL.DblBundle
 	/// itself cannot be abstract because in the event of a loading error, we need to deserialize it
 	/// to properly report the situation to the user.
 	/// </summary>
-	public abstract class DblMetadataBase<TL> : DblMetadata, IProjectInfo where TL: DblMetadataLanguage, new()
+	public abstract class DblMetadataBase<TL> : DblMetadata, IProjectInfo
+		where TL: DblMetadataLanguage, new()
 	{
 		/// <summary>
 		/// Loads information about a Digital Bible Library bundle from the specified reader.
 		/// </summary>
-		/// <param name="metadataReader">A TextReader object assumed to be positioned at the start of the </param>
-		/// <param name="resourceIdentifier">Filename or other meaningful identifier that can be used in an error message if there is a
-		/// problem reading the data or creating the desired type of DblMetadata object from it.</param>
-		/// <param name="exception">Set if the metadata object could not be instantiated or initialized.</param>
-		public static T Load<T>(TextReader metadataReader, string resourceIdentifier,  out Exception exception) where T : DblMetadataBase<TL>
+		/// <param name="metadataReader">A TextReader object assumed to be positioned at the
+		/// start of the metadata (leading XML element is permitted, but not required)</param>
+		/// <param name="resourceIdentifier">Filename or other meaningful identifier that can be
+		/// used in an error message if there is a
+		/// problem reading the data or creating the desired type of DblMetadata object from it.
+		/// </param>
+		/// <param name="exception">Set if the metadata object could not be instantiated or
+		/// initialized.</param>
+		public static T Load<T>(TextReader metadataReader, string resourceIdentifier,
+			out Exception exception) where T : DblMetadataBase<TL>
 		{
 			string data;
 			try
@@ -77,7 +84,11 @@ namespace SIL.DblBundle
 			if (metadata == null)
 			{
 				if (exception == null)
-					exception = new ApplicationException($"Loading metadata ({resourceIdentifier}) was unsuccessful.");
+				{
+					exception = new ApplicationException(
+						$"Loading metadata ({resourceIdentifier}) was unsuccessful.");
+				}
+
 				return null;
 			}
 			try
@@ -92,12 +103,21 @@ namespace SIL.DblBundle
 		}
 
 		/// <summary>
-		/// Loads information about a Digital Bible Library bundle from the specified projectFilePath.
+		/// Loads information about a Digital Bible Library bundle from the specified filePath.
 		/// </summary>
-		public static T Load<T>(string projectFilePath, out Exception exception) where T : DblMetadataBase<TL>
+		public static T Load<T>(string filePath, out Exception exception)
+			where T : DblMetadataBase<TL>
 		{
-			using (var stream = new FileStream(projectFilePath, FileMode.Open))
-				return Load<T>(new StreamReader(stream), projectFilePath, out exception);
+			try
+			{
+				using (var stream = new FileStream(filePath, FileMode.Open))
+					return Load<T>(new StreamReader(stream), filePath, out exception);
+			}
+			catch (Exception e)
+			{
+				exception = e;
+				return null;
+			}
 		}
 
 		/// <summary>