+++ CHANGELOG.md
@@ -17,7 +17,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ## [Unreleased]
 
 ### Added
-- DblMetadata.Load overload to allow deserialization from a TextReader.
+- [SIL.DblBundle] DblMetadata.Load overload to allow deserialization from a TextReader.
 
 ### Changed